feat: refactor cmake (#26)
* feat:  Initial refactor of cmake

Refactor cmake and project organization to be more modular and to better conform to best practices.

This commit is to test CI before moving on to PlatformIO.

* fix:  Update CI scripts for test path

* feat: Refactor cmake

Refactored cmake to make the project easier to include in end-user applications.  Updated PlatformIO as needed.  Removed usage of monolithic header in tests to simplify PIO compiles and reduce compliation unit size.

* fix:  Correct back slash to forward slash