fix: rename helpers
Cpp-Crypto uses the header 'helpers/helpers.h'.
Cpp-Client also uses this 'helpers/helpers.h' naming.
These headers are not identical and contain only “helper” functions specific to its library.

When both Cpp-Client & Cpp-Crypto are included in the same project, it results in a name collision causing one of the headers to be dropped.

This PR renames 'helpers/helpers.h' to 'helpers/crypto_helpers.h’, and resolves the name collision.